backend/requirements.txt
fastapi[standard]
sqlalchemy>=2.0
psycopg[binary]>=3.1
pydantic-settings>=2.0
argon2-cffi